log_reader finished

+++ client/header/data.h
@@ -11,7 +11,7 @@ struct LogRec
 {
     char log_name[32];
     pid_t pid;
-    long log_time;
+    int log_time;
     char log_ip[32];
     short log_type;
 };
@@ -20,9 +20,9 @@ struct MatchedLogRec
 {
     char log_name[32];
     pid_t pid;
-    long login_time;
-    long logout_time;
-    long duration;
+    int  login_time;
+    int  logout_time;
+    int  duration;
     char log_ip[32];
     friend ifstream & operator>>(ifstream & fin,MatchedLogRec & matched_log);
     friend ofstream & operator<<(ofstream & fout,MatchedLogRec const & matched_log);
